Book-Flask-Backend-220107/chapter09/requirements.txt
attrs==21.4.0
Automat==20.2.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
constantly==15.1.0
Flask==2.0.2
Flask-Script==2.0.6
Flask-Twisted==0.1.2
greenlet==1.1.2
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
observable==1.0.3
packaging==21.3
pluggy==1.0.0
protobuf==3.19.3
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pytest==6.2.5
six==1.16.0
SQLAlchemy==1.4.29
toml==0.10.2
Twisted==21.7.0
typing_extensions==4.0.1
Werkzeug==2.0.2
zope.interface==5.4.0
